requirements.txt
requests
sentinelsat
oauthlib
requests_oauthlib
python-dotenv
colorlog
